# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "What Deck API",
        "description": "Magic: The Gathering card collection management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/card-instances": {
            "get": {
                "tags": [
                    "Card Instances"
                ],
                "summary": "List all card instances",
                "description": "Get a paginated list of card instances with optional filtering",
                "operationId": "43cb51bad0c2592e94aeb9b6058ec63e",
                "parameters": [
                    {
                        "name": "collection_id",
                        "in": "query",
                        "description": "Filter by collection ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "deck_id",
                        "in": "query",
                        "description": "Filter by deck ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Filter available instances (not in deck)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CardInstance"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Card Instances"
                ],
                "summary": "Create a new card instance",
                "description": "Create a new physical card instance in a collection",
                "operationId": "20316a63c8e300e426caed664c478b44",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_id",
                                    "collection_id"
                                ],
                                "properties": {
                                    "card_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "collection_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "condition": {
                                        "type": "string",
                                        "enum": [
                                            "mint",
                                            "near_mint",
                                            "lightly_played",
                                            "moderately_played",
                                            "heavily_played",
                                            "damaged"
                                        ],
                                        "example": "near_mint"
                                    },
                                    "foil": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Card instance created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardInstance"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/card-instances/{id}": {
            "get": {
                "tags": [
                    "Card Instances"
                ],
                "summary": "Get a specific card instance",
                "description": "Retrieve a specific card instance with its relationships",
                "operationId": "b76383f8f51b4cd598f3c8098e222af0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card Instance ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardInstance"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card instance not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Card Instances"
                ],
                "summary": "Update a card instance",
                "description": "Update an existing card instance (condition, foil status)",
                "operationId": "6bebdc34c1aa1897d2754977edfe3fa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card Instance ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "condition": {
                                        "type": "string",
                                        "enum": [
                                            "mint",
                                            "near_mint",
                                            "lightly_played",
                                            "moderately_played",
                                            "heavily_played",
                                            "damaged"
                                        ],
                                        "example": "lightly_played"
                                    },
                                    "foil": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Card instance updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardInstance"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card instance not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Card Instances"
                ],
                "summary": "Delete a card instance",
                "description": "Delete a physical card instance",
                "operationId": "3805861d14d0653e8969dd247ba08888",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card Instance ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card instance deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card instance not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/card-instances/{cardInstanceId}/move-to-deck/{deckId}": {
            "put": {
                "tags": [
                    "Card Instances"
                ],
                "summary": "Move card instance to deck",
                "description": "Move a card instance to a specific deck",
                "operationId": "35d85c3b86b00151ec0a2c9adc87eafb",
                "parameters": [
                    {
                        "name": "cardInstanceId",
                        "in": "path",
                        "description": "Card Instance ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "deckId",
                        "in": "path",
                        "description": "Deck ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card instance moved to deck",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardInstance"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Deck and collection must belong to same user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card instance or deck not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/card-instances/{id}/remove-from-deck": {
            "put": {
                "tags": [
                    "Card Instances"
                ],
                "summary": "Remove card instance from deck",
                "description": "Remove a card instance from its current deck",
                "operationId": "a255686d5ca4b47939becb9955060fe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card Instance ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card instance removed from deck",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardInstance"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card instance not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "title": "Paginated Response",
                "description": "Standard paginated response format",
                "properties": {
                    "data": {
                        "description": "Array of resource items",
                        "type": "array",
                        "items": {}
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 5
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 67
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=1"
                            },
                            "last": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=5"
                            },
                            "prev": {
                                "type": "string",
                                "format": "url",
                                "example": null,
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=2",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "SuccessResponse": {
                "title": "Success Response",
                "description": "Standard success response format",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operation completed successfully"
                    }
                },
                "type": "object"
            },
            "CardInstance": {
                "title": "Card Instance Resource",
                "description": "Individual physical card instance resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "card_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "collection_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "deck_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "condition": {
                        "type": "string",
                        "enum": [
                            "mint",
                            "near_mint",
                            "lightly_played",
                            "moderately_played",
                            "heavily_played",
                            "damaged"
                        ],
                        "example": "near_mint"
                    },
                    "foil": {
                        "type": "boolean",
                        "example": false
                    },
                    "language": {
                        "type": "string",
                        "example": "English"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "red",
                            "burn",
                            "instant"
                        ],
                        "nullable": true
                    },
                    "purchase_price": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5,
                        "nullable": true
                    },
                    "alter": {
                        "type": "boolean",
                        "example": false
                    },
                    "proxy": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "card": {
                        "$ref": "#/components/schemas/Card"
                    },
                    "collection": {
                        "$ref": "#/components/schemas/Collection"
                    },
                    "deck": {
                        "$ref": "#/components/schemas/Deck"
                    },
                    "is_in_deck": {
                        "description": "Whether this instance is assigned to a deck",
                        "type": "boolean",
                        "example": true
                    },
                    "is_available": {
                        "description": "Whether this instance is available (not in a deck)",
                        "type": "boolean",
                        "example": false
                    },
                    "condition_description": {
                        "description": "Human-readable condition description",
                        "type": "string",
                        "example": "Near Mint (NM)"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}